Incremental updates

+++ src/app/orders/order-detail/order-detail.page.ts
@@ -3,20 +3,15 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import {
   AlertController,
-  ModalController,
   NavController,
   ToastController,
 } from '@ionic/angular';
 import { Customer } from 'src/app/classes/customer.model';
-import { Item } from 'src/app/classes/item.model';
-import { Menu } from 'src/app/classes/menu.model';
-import { OrderDto } from 'src/app/classes/order-dto.model';
 import { OrderMenuDto } from 'src/app/classes/order-menu-dto.model';
 import { OrderMenuIngredient } from 'src/app/classes/order-menu-ingredient.models';
 import { OrderMenu } from 'src/app/classes/order-menu.model';
 import { Order } from 'src/app/classes/order.model';
-import { Uom } from 'src/app/classes/uom.model';
-import { CustomerSearchComponent } from 'src/app/customers/customer-search/customer-search.component';
+import { OrderDetailsConfig } from 'src/app/Configurations/order-details.config';
 import { OrdersService } from 'src/app/services/orders.service';
 
 @Component({
@@ -28,8 +23,7 @@ export class OrderDetailPage implements OnInit {
   orderMenus: OrderMenuDto[] = [];
   customer = new Customer();
   order = new Order();
-  orderStatus = 0;
-  orderStatusColor: string;
+  orderDetailsConfig = new OrderDetailsConfig();
 
   isFetching = false;
 
@@ -60,22 +54,7 @@ export class OrderDetailPage implements OnInit {
       this.orderService.getOrder(orderId).subscribe((resData) => {
         this.order = resData;
         this.customer = resData.customer;
-        if (this.order.orderStatus === 'Preparing') {
-          this.orderStatus = 1;
-          this.orderStatusColor = 'warning';
-        }
-        if (this.order.orderStatus === 'In Transit') {
-          this.orderStatus = 2;
-          this.orderStatusColor = 'tertiary';
-        }
-        if (this.order.orderStatus === 'Delivered') {
-          this.orderStatus = 3;
-          this.orderStatusColor = 'success';
-        }
-        if (this.order.orderStatus === 'Cancelled') {
-          this.orderStatus = 4;
-          this.orderStatusColor = 'primary';
-        }
+        this.orderDetailsConfig.setParams(this.order.orderStatus);
       });
 
       this.orderService.getOrderMenus(orderId).subscribe((resData) => {
@@ -85,30 +64,42 @@ export class OrderDetailPage implements OnInit {
     });
   }
 
+  onUpdateStatus(newStatus: string) {
+    this.order.orderStatus = newStatus;
+    this.orderService.patchOrders(this.order).subscribe(
+      res => this.orderService.ordersHaveChanged.next(true)
+    );
+    this.orderDetailsConfig.setParams(newStatus);
+
+  }
+
   onDeleteMenu(menu: OrderMenu) {
-    this.alertCtrl
-      .create({
-        header: 'Confirm',
-        message: 'This will be deleted permanently .',
-        buttons: [
-          {
-            text: 'Cancel',
-          },
-          {
-            text: 'Delete',
-            handler: () => {
-              this.orderService
-                .deleteOrderMenu(menu.orderMenuId)
-                .subscribe((res) => {
-                  this.removeMenuObj(menu);
-                });
+    if (this.orderDetailsConfig.deleteIngredientButton) {
+      this.alertCtrl
+        .create({
+          header: 'Confirm',
+          message: 'This will be deleted permanently .',
+          buttons: [
+            {
+              text: 'Cancel',
             },
-          },
-        ],
-      })
-      .then((res) => {
-        res.present();
-      });
+            {
+              text: 'Delete',
+              handler: () => {
+                this.orderService
+                  .deleteOrderMenu(menu.orderMenuId)
+                  .subscribe((res) => {
+                    this.removeMenuObj(menu);
+                  });
+              },
+            },
+          ],
+        })
+        .then((res) => {
+          res.present();
+        });
+    }
+
   }
 
   removeMenuObj(menu: OrderMenu) {
@@ -120,7 +111,7 @@ export class OrderDetailPage implements OnInit {
   }
 
   onDeleteIngredient(ing: OrderMenuIngredient, ings: OrderMenuIngredient[]) {
-    if (this.orderStatus < 2) {
+    if (this.orderDetailsConfig.deleteIngredientButton) {
       this.alertCtrl
         .create({
           header: 'Confirm',
